python-langchain/requirements.txt
langchain
langchain_openai
langchain_community
langchain_core
python-dotenv
langgraph
numexpr
pypdf
unstructured
psutil
langchain-chroma
chromadb
